Code/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0

# Deep learning utilities
einops>=0.6.0

# Image processing
scikit-image>=0.19.0
Pillow>=9.0.0

# Visualization
matplotlib>=3.5.0

# Data handling
pandas>=1.4.0
pyyaml>=6.0
tqdm>=4.62.0

# Medical imaging
pydicom>=2.3.0
nibabel>=4.0.0

# Optional: PyRoNN for CT projection
# Install from: https://github.com/csyben/PYRO-NN
# pyronn

# Development
pytest>=7.0.0
black>=22.0.0
